requirements.txt
gradio>=3.50.2
pyyaml>=6.0
swanlab 